store system/customer/requirements.txt
flask
flask_sqlalchemy
PyJWT
mysql-connector-python
py-solc-x
web3